Upgraded RestExpress-Common dependency to 1.0.2, moving portions of Query…sTest to RestExpress-Common also.  Changed Response.setCollectionResponse() to use the newly-cloneable QueryRange.
